refactor: restructured the repository
