+++ .github/workflows/test.yml
@@ -14,7 +14,7 @@ jobs:
               uses: actions/setup-go@v5
 
             - name: Build
-              run: go build ./...
+              run: go build main.go
     
     golangci:
         runs-on: ubuntu-latest
@@ -43,4 +43,4 @@ jobs:
               uses: actions/setup-go@v5
 
             - name: Test
-              run: go test ./...
+              run: go test ./test
