+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/codescalersinternships/Datetime-Client-Fatma-Ebrahim
+
+go 1.23.1
+
+require github.com/cenkalti/backoff/v4 v4.3.0
